chore: clean up unused scripts and re-organize modules
