+++ src/scratch/_sketch.py
